correct the default path
